# openapi.yaml
openapi: 3.0.3
info:
  title: Conference Tiebreaker API
  version: 1.0.0
  description: API for fetching games, standings, and simulating conference tiebreaker scenarios
servers:
  - url: /api
    description: Local development server
paths:
...
  /cfbd-alert-handler:
    post:
      summary: Webhook handler for CFBD API usage alerts
      operationId: postCfbdAlertHandler
      tags:
        - Monitoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CfbdAlertRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CfbdAlertResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CfbdAlertRequest:
      type: object
      required:
        - remainingCalls
        - patronLevel
        - threshold
      properties:
        remainingCalls:
          type: integer
          format: int32
        patronLevel:
          type: integer
          format: int32
        threshold:
          type: integer
          format: int32
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
          nullable: true
    CfbdAlertResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        method:
          type: string
          nullable: true
        emailId:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...